move inline wrapper component to wrapper component
